fixed bug in deleteCourse that caused crash, ui and usability improvements, cleaned up project

+++ src/main/java/com/example/ischedule/Repository/CourseScheduleRepository.java
@@ -9,7 +9,7 @@ import java.sql.Time;
 import java.util.List;
 
 @Repository
-public interface CourseScheduleRepository extends JpaRepository<CourseSchedule, Long> {
+public interface CourseScheduleRepository extends JpaRepository<CourseSchedule, Integer> {
     /*
      * Find all course schedules for a specific course.
      */
